wip: merging detox-server and detox-common into detox package
