--- detox-server/.gitignore
@@ -1,6 +1,0 @@
-node_modules
-*.log
-.DS_Store
-coverage
-.idea
-*.iml
