--- detox-server/README.md
@@ -1,14 +1,0 @@
-> detox
-
-# Server
-
-## Install
-
-* make sure you're in folder `detox/detox-server`
-* run `npm install`
-
-## Run the server
-
-* make sure you're in folder `detox/detox-server`
-* run `npm start`
-* you should see `server listening on localhost:8099...`
