--- detox-server/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "detox-server",
-  "description": "Websocket proxy server for detox",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
-  },
-  "version": "7.0.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/wix/detox.git"
-  },
-  "bin": {
-    "detox-server": "src/cli.js"
-  },
-  "scripts": {
-    "start": "node src/cli.js",
-    "test": ":"
-  },
-  "bugs": {
-    "url": "https://github.com/wix/detox/issues"
-  },
-  "homepage": "https://github.com/wix/detox#readme",
-  "main": "./src/index.js",
-  "author": "Tal Kol <talkol@gmail.com>",
-  "license": "MIT",
-  "dependencies": {
-    "detox-common": "^8.0.0",
-    "lodash": "^4.13.1",
-    "ws": "^1.1.0"
-  }
-}
