--- detox-server/src/cli.js
@@ -1,4 +1,0 @@
-#! /usr/bin/env node
-const DetoxServer = require('./DetoxServer');
-
-const detoxServer = new DetoxServer(8099);
