--- detox-server/src/index.js
@@ -1,3 +1,0 @@
-const DetoxServer = require('./DetoxServer');
-
-module.exports = DetoxServer;
