+++ detox/package.json
@@ -38,10 +38,10 @@
     "prettier": "1.7.0"
   },
   "dependencies": {
+    "bunyan": "^1.8.12",
+    "bunyan-debug-stream": "noomorph/bunyan-debug-stream",
     "child-process-promise": "^2.2.0",
     "commander": "^2.15.1",
-    "detox-common": "^8.0.0",
-    "detox-server": "^7.0.0",
     "fs-extra": "^4.0.2",
     "get-port": "^2.1.0",
     "ini": "^1.3.4",
@@ -67,12 +67,14 @@
       "src/artifacts/log",
       "src/artifacts/screenshot",
       "src/artifacts/video",
+      "src/server/DetoxServer.js",
       ".*Driver.js",
       "EmulatorTelnet.js",
       "Emulator.js",
       "DeviceDriverBase.js",
       "GREYConfiguration.js",
       "src/utils/environment.js",
+      "src/utils/logger.js",
       "src/utils/onTerminate.js",
       "src/utils/sleep.js",
       "AAPT.js",
