+++ detox/src/server/DetoxServer.js
@@ -1,13 +1,12 @@
 const _ = require('lodash');
 const WebSocketServer = require('ws').Server;
-const log = require('detox-common').logger.server.child({ __filename });
 
 class DetoxServer {
-  constructor(port) {
+  constructor({ port, log }) {
     this.wss = new WebSocketServer({ port });
     this.sessions = {};
-
-    log.info(`server listening on localhost:${this.wss.options.port}...`);
+    this.log = log.child({ __filename });
+    this.log.info(`server listening on localhost:${this.wss.options.port}...`);
     this._setup();
   }
 
@@ -25,24 +24,24 @@ class DetoxServer {
             if (action.params && action.params.sessionId && action.params.role) {
               sessionId = action.params.sessionId;
               role = action.params.role;
-              log.debug({ event: 'LOGIN' }, `role=${role}, sessionId=${sessionId}`);
+              this.log.debug({ event: 'LOGIN' }, `role=${role}, sessionId=${sessionId}`);
               _.set(this.sessions, [sessionId, role], ws);
               action.type = 'loginSuccess';
               this.sendAction(ws, action);
-              log.debug({ event: 'LOGIN_SUCCESS' }, `role=${role}, sessionId=${sessionId}`);
+              this.log.debug({ event: 'LOGIN_SUCCESS' }, `role=${role}, sessionId=${sessionId}`);
             }
           } else if (sessionId && role) {
-            log.trace({ event: 'MESSAGE', action: action.type }, `role=${role} action=${action.type} (sessionId=${sessionId})`);
+            this.log.trace({ event: 'MESSAGE', action: action.type }, `role=${role} action=${action.type} (sessionId=${sessionId})`);
             this.sendToOtherRole(sessionId, role, action);
           }
         } catch (err) {
-          log.debug({ event: 'ERROR', err }, `Invalid JSON received, cannot parse`, err);
+          this.log.debug({ event: 'ERROR', err }, `Invalid JSON received, cannot parse`, err);
         }
       });
 
       ws.on('close', () => {
         if (sessionId && role) {
-          log.debug({ event: 'DISCONNECT' }, `role=${role}, sessionId=${sessionId}`);
+          this.log.debug({ event: 'DISCONNECT' }, `role=${role}, sessionId=${sessionId}`);
           _.set(this.sessions, [sessionId, role], undefined);
         }
       });
@@ -61,7 +60,7 @@ class DetoxServer {
     if (ws) {
       this.sendAction(ws, action);
     } else {
-      log.debug({ event: 'CANNOT_FORWARD' }, `role=${otherRole} not connected, cannot fw action (sessionId=${sessionId})`);
+      this.log.debug({ event: 'CANNOT_FORWARD' }, `role=${otherRole} not connected, cannot fw action (sessionId=${sessionId})`);
     }
   }
